♻️ fix: refactor Card
