+++ client/app/(user)/ui-demo.tsx
@@ -7,7 +7,7 @@ import { View } from '@/components/ui/View';
 import { Text } from '@/components/ui/Text';
 import { Button } from '@/components/ui/Button/Button';
 import { TextInput } from '@/components/ui/Input';
-import { Card } from '@/components/ui/Card';
+import { Card } from '@/components/ui/Card/Card';
 import { Stack, Inline, Section } from '@/components/ui/Layout';
 import { Header } from '@/components/ui/Header';
